requirements.txt
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.44.1
streamlit==1.31.0
google-generativeai==0.4.1
python-dotenv==1.0.1
scipy==1.12.0
joblib==1.3.2
